1927-sum-game Medium .cpp Math
Solved by: vaibhav_jpd 
LinkedIn: https://www.linkedin.com/in/vaibhav-raj-singh-/

LeetCode: leetcode.com/u/rosvert/

Portfolio:
https://www.vaibhavrajsingh.in/

#vaibhav-raj-singh
#jiit
#Jaypee

Student @Jaypee Institute of Information Technology

+++ C++/Medium/1927-sum-game/1927-sum-game.cpp
@@ -1,16 +1,16 @@
 class Solution {
 public:
     bool sumGame(string s) {
-        double res = 0;
-        double n = s.length();
+        int res = 0;
+        int n = s.length();
         for (int i = 0; i < n; i++) {
-            double sign;
+            int sign;
             if (i < n / 2) {
                 sign = 2;
             } else {
                 sign = -2;
             }
-            double value;
+            int value;
             if (s[i] == '?') {
                 value = 9;
                 res -= 9 * sign/2;
